+++ sdk/ai/azure-ai-agents/azure/ai/agents/aio/operations/_patch.py
@@ -24,6 +24,7 @@ from typing import (
     Dict,
     List,
     Optional,
+    Type,
     Union,
     cast,
     overload,
@@ -442,7 +443,7 @@ class RunsOperations(RunsOperationsGenerated):
         response_format: Optional["_types.AgentsResponseFormatOption"] = None,
         parallel_tool_calls: Optional[bool] = None,
         metadata: Optional[Dict[str, str]] = None,
-        run_handler: Optional[_models.AsyncRunHandler] = None,
+        run_handler: _models.AsyncRunHandler | Type[_models.AsyncRunHandler] = _models.AsyncRunHandler,
         polling_interval: int = 1,
         **kwargs: Any,
     ) -> _models.ThreadRun:
@@ -525,7 +526,7 @@ class RunsOperations(RunsOperationsGenerated):
         :paramtype metadata: dict[str, str]
         :keyword run_handler: Optional handler to customize run processing and tool execution.
             Default value is None.
-        :paramtype run_handler: ~azure.ai.agents.models.AsyncRunHandler
+        :paramtype run_handler: Union[~azure.ai.agents.models.AsyncRunHandler, Type[~azure.ai.agents.models.AsyncRunHandler]]
         :keyword polling_interval: The time in seconds to wait between polling the service for run status.
             Default value is 1.
         :paramtype polling_interval: int
@@ -557,10 +558,9 @@ class RunsOperations(RunsOperationsGenerated):
         )
 
         # Monitor and process the run status
-        if not run_handler:
-            run_handler = _models.AsyncRunHandler()
+        run_handler_obj = run_handler() if isinstance(run_handler, type) else run_handler
 
-        return await run_handler._start(self, run, polling_interval)  # pylint: disable=protected-access
+        return await run_handler_obj._start(self, run, polling_interval)  # pylint: disable=protected-access
 
     @overload
     async def stream(