+++ sdk/ai/azure-ai-agents/azure/ai/agents/operations/_patch.py
@@ -23,6 +23,7 @@ from typing import (
     Iterator,
     List,
     Optional,
+    Type,
     Union,
     cast,
     overload,
@@ -441,7 +442,7 @@ class RunsOperations(RunsOperationsGenerated):
         response_format: Optional["_types.AgentsResponseFormatOption"] = None,
         parallel_tool_calls: Optional[bool] = None,
         metadata: Optional[Dict[str, str]] = None,
-        run_handler: Optional[_models.RunHandler] = None,
+        run_handler: _models.RunHandler | Type[_models.RunHandler] = _models.RunHandler,
         polling_interval: int = 1,
         **kwargs: Any,
     ) -> _models.ThreadRun:
@@ -524,7 +525,7 @@ class RunsOperations(RunsOperationsGenerated):
         :paramtype metadata: dict[str, str]
         :keyword run_handler: Optional handler to customize run processing and tool execution.
             Default value is None.
-        :paramtype run_handler: ~azure.ai.agents.models.RunHandler
+        :paramtype run_handler: Union[~azure.ai.agents.models.RunHandler, Type[~azure.ai.agents.models.RunHandler]]
         :keyword polling_interval: The time in seconds to wait between polling the service for run status.
             Default value is 1.
         :paramtype polling_interval: int
@@ -556,10 +557,9 @@ class RunsOperations(RunsOperationsGenerated):
         )
 
         # Monitor and process the run status
-        if not run_handler:
-            run_handler = _models.RunHandler()
+        run_handler_obj = run_handler() if isinstance(run_handler, type) else run_handler
 
-        return run_handler._start(self, run, polling_interval)  # pylint: disable=protected-access
+        return run_handler_obj._start(self, run, polling_interval)  # pylint: disable=protected-access
 
     @overload
     def stream(